pyproject.toml
[project]
name = "daylio-data-cleaner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
